requirements.txt
# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Data Collection
yfinance>=0.2.28

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Dashboard
streamlit>=1.28.0

# Configuration
pyyaml>=6.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Optional: Deep Learning
torch>=2.0.0
tensorflow>=2.13.0

# Development Tools
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0